pyproject.toml
[project]
name = "finegrain_toolbox"
version = "0.1.0"
description = "Utilities for Finegrain models."
authors = [
    { name = "Pierre Chapuis", email = "git@catwell.info" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.8.0",
    "diffusers>=0.35.1",
    "transformers>=4.56.0",
    "accelerate>=1.10.1",
    "numpy>=1.26.4",
    "pillow>=11.3.0",
    "peft>=0.17.1",
    "safetensors>=0.6.2",
]

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
src = ["src"]  # https://docs.astral.sh/ruff/settings/#src
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "UP",  # pyupgrade
  "A",   # flake8-builtins
  "B",   # flake8-bugbear
  "Q",   # flake8-quotes
  "I",   # isort
]

[tool.ruff.lint.isort]
section-order = ["future", "standard-library", "third-party", "finegrain_toolbox", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
finegrain_toolbox = ["finegrain_toolbox"]

[tool.pyright]
include = ["src/finegrain_toolbox"]
exclude = ["**/__pycache__"]
strict = ["."]

[dependency-groups]
dev = [
    "pyright>=1.1.404",
    "ruff>=0.12.11",
]
